refactor(estimates): extract EstimateTierEditor shared component (Phase 2c/2d)
Inline multi-option tier card grid extracted from NewEstimateForm into
EstimateTierEditor. EstimateEditForm TierEditor left as-is (self-contained
form component with own state + API, not a duplicated UI section).
