+++ apps/web/app/app/estimates/new/NewEstimateForm.tsx
@@ -15,6 +15,7 @@ import { PREP_LEVEL_MULTIPLIERS, getMaterialsByCategory } from "@ai-fsm/domain";
 import { formatCents } from "@/lib/estimates/pricing";
 import { GuardrailsSection } from "../components/GuardrailsSection";
 import { LineItemsTable } from "../components/LineItemsTable";
+import { EstimateTierEditor } from "../components/EstimateTierEditor";
 import { InlineClientForm } from "./InlineClientForm";
 import { InlineJobForm } from "./InlineJobForm";
 import { InlinePropertyForm } from "./InlinePropertyForm";
@@ -786,132 +787,16 @@ export function NewEstimateForm(props: NewEstimateFormProps) {
                     />
                   </div>
                 ) : mode === "multi_option" ? (
-                  <div style={{ display: "grid", gridTemplateColumns: "repeat(3, 1fr)", gap: "var(--space-4)", marginBottom: "var(--space-3)" }}>
-                    {tiers.map((tier, ti) => {
-                      const tierSub = tierSubtotalCents(tier);
-                      const tierTax = Math.round((tierSub * taxRateNum) / 100);
-                      const tierTotal = tierSub + tierTax;
-                      return (
-                        <Card key={ti} padding="sm" style={{
-                          border: tier.is_recommended ? "2px solid var(--accent)" : "1px solid var(--border)",
-                          position: "relative",
-                        }}>
-                          {tier.is_recommended && (
-                            <div style={{
-                              position: "absolute", top: -10, left: "50%", transform: "translateX(-50%)",
-                              background: "var(--accent)", color: "#fff", padding: "2px 12px", borderRadius: 99,
-                              fontSize: "var(--text-xs)", fontWeight: 600, whiteSpace: "nowrap",
-                            }}>
-                              Recommended
-                            </div>
-                          )}
-                          <div style={{ marginBottom: "var(--space-2)" }}>
-                            <input
-                              className="p7-input"
-                              type="text"
-                              value={tier.label}
-                              onChange={(e) => updateTier(ti, { label: e.target.value })}
-                              placeholder="Option label"
-                              disabled={pending}
-                              style={{ fontWeight: 700, fontSize: "var(--text-lg)", marginBottom: "var(--space-1)" }}
-                            />
-                            <input
-                              className="p7-input"
-                              type="text"
-                              value={tier.description}
-                              onChange={(e) => updateTier(ti, { description: e.target.value })}
-                              placeholder="Brief description"
-                              disabled={pending}
-                              style={{ fontSize: "var(--text-sm)", color: "var(--fg-muted)" }}
-                            />
-                          </div>
-
-                          <label style={{ display: "flex", alignItems: "center", gap: "var(--space-1)", cursor: "pointer", marginBottom: "var(--space-2)", fontSize: "var(--text-sm)" }}>
-                            <input
-                              type="checkbox"
-                              checked={tier.is_recommended}
-                              onChange={(e) => updateTier(ti, { is_recommended: e.target.checked })}
-                              disabled={pending}
-                            />
-                            <span>Mark as recommended</span>
-                          </label>
-
-                          <div style={{ borderTop: "1px solid var(--border)", paddingTop: "var(--space-2)", marginTop: "var(--space-2)" }}>
-                            {tier.line_items.map((row, li) => (
-                              <div key={li} style={{ marginBottom: "var(--space-2)" }}>
-                                <input
-                                  className="p7-input"
-                                  type="text"
-                                  value={row.description}
-                                  onChange={(e) => updateTierLineItem(ti, li, "description", e.target.value)}
-                                  placeholder="Description"
-                                  disabled={pending}
-                                  style={{ marginBottom: "var(--space-1)", fontSize: "var(--text-sm)" }}
-                                />
-                                <div style={{ display: "flex", gap: "var(--space-2)" }}>
-                                  <input
-                                    className="p7-input"
-                                    type="number"
-                                    min="0.01"
-                                    step="0.01"
-                                    value={row.quantity}
-                                    onChange={(e) => updateTierLineItem(ti, li, "quantity", e.target.value)}
-                                    disabled={pending}
-                                    style={{ width: 60, fontSize: "var(--text-sm)" }}
-                                  />
-                                  <input
-                                    className="p7-input"
-                                    type="number"
-                                    min="0"
-                                    step="0.01"
-                                    value={row.unit_price}
-                                    onChange={(e) => updateTierLineItem(ti, li, "unit_price", e.target.value)}
-                                    disabled={pending}
-                                    style={{ width: 90, fontSize: "var(--text-sm)" }}
-                                  />
-                                  <span style={{ fontSize: "var(--text-sm)", color: "var(--fg-muted)", lineHeight: "var(--space-5)" }}>
-                                    {formatCents(lineTotal(row))}
-                                  </span>
-                                  {tier.line_items.length > 1 && (
-                                    <button
-                                      type="button"
-                                      onClick={() => removeTierLineItem(ti, li)}
-                                      disabled={pending}
-                                      style={{ background: "none", border: "none", cursor: "pointer", color: "var(--color-danger)", fontSize: "var(--text-sm)", padding: 0, lineHeight: 1 }}
-                                      aria-label="Remove line item"
-                                    >
-                                      ×
-                                    </button>
-                                  )}
-                                </div>
-                              </div>
-                            ))}
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="sm"
-                              onClick={() => addTierLineItem(ti)}
-                              disabled={pending}
-                              style={{ width: "100%", marginTop: "var(--space-1)" }}
-                            >
-                              + Add item
-                            </Button>
-                          </div>
-
-                          <div style={{ borderTop: "1px solid var(--border)", paddingTop: "var(--space-2)", marginTop: "var(--space-3)", textAlign: "right" }}>
-                            {tierTax > 0 && (
-                              <div style={{ fontSize: "var(--text-sm)", color: "var(--fg-muted)" }}>
-                                Tax: {formatCents(tierTax)}
-                              </div>
-                            )}
-                            <div style={{ fontSize: "var(--text-lg)", fontWeight: 700 }}>
-                              {formatCents(tierTotal)}
-                            </div>
-                          </div>
-                        </Card>
-                      );
-                    })}
-                  </div>
+                  <EstimateTierEditor
+                    tiers={tiers}
+                    taxRateNum={taxRateNum}
+                    disabled={pending}
+                    onUpdateTier={updateTier}
+                    onUpdateTierLineItem={updateTierLineItem}
+                    onAddTierLineItem={addTierLineItem}
+                    onRemoveTierLineItem={removeTierLineItem}
+                    tierSubtotalCents={tierSubtotalCents}
+                  />
                 ) : (
                   <>
                     {/* Item Suggester */}
